+++ ext/opentelemetry-ext-psycopg2/setup.cfg
@@ -1,4 +1,4 @@
-# Copyright 2019, OpenTelemetry Authors
+# Copyright 2020, OpenTelemetry Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,13 +13,13 @@
 # limitations under the License.
 #
 [metadata]
-name = opentelemetry-ext-postgresql
-description = OpenTelemetry PostgreSQL integration
+name = opentelemetry-ext-psycopg2
+description = OpenTelemetry psycopg2 integration
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 author = OpenTelemetry Authors
 author_email = cncf-opentelemetry-contributors@lists.cncf.io
-url = https://github.com/open-telemetry/opentelemetry-python/ext/opentelemetry-ext-postgresql
+url = https://github.com/open-telemetry/opentelemetry-python/ext/opentelemetry-ext-psycopg2
 platforms = any
 license = Apache-2.0
 classifiers =
@@ -39,7 +39,7 @@ package_dir=
     =src
 packages=find_namespace:
 install_requires =
-    opentelemetry-api >= 0.3.dev0
+    opentelemetry-api >= 0.4.dev0
     psycopg2-binary >= 2.7.3.1
     wrapt >= 1.0.0, < 2.0.0
